Working on views for each phone
This requires that I make sure to look at if the Height and Weight are either compact or regular and making sure the phones without those views have to be adjusted for. Also make sure we set pin constraints so that views with the same views adjust for their own, or things get really fucky.